+++ package.json
@@ -1,96 +1,96 @@
 {
-  "name": "playcanvas", 
-  "version": "1.44.0-dev", 
-  "author": "PlayCanvas <support@playcanvas.com>", 
-  "homepage": "https://playcanvas.com", 
-  "description": "PlayCanvas WebGL game engine", 
+  "name": "playcanvas",
+  "version": "1.44.0-dev",
+  "author": "PlayCanvas <support@playcanvas.com>",
+  "homepage": "https://playcanvas.com",
+  "description": "PlayCanvas WebGL game engine",
   "keywords": [
-    "3D", 
-    "2D", 
-    "VR", 
-    "WebGL", 
-    "WebGL2", 
-    "game", 
-    "engine", 
-    "HTML5", 
-    "browser", 
+    "3D",
+    "2D",
+    "VR",
+    "WebGL",
+    "WebGL2",
+    "game",
+    "engine",
+    "HTML5",
+    "browser",
     "typings"
-  ], 
-  "license": "MIT", 
-  "main": "build/playcanvas.js", 
-  "module": "build/playcanvas.mjs", 
-  "types": "build/playcanvas.d.ts", 
+  ],
+  "license": "MIT",
+  "main": "build/playcanvas.js",
+  "module": "build/playcanvas.mjs",
+  "types": "build/playcanvas.d.ts",
   "bugs": {
     "url": "https://github.com/playcanvas/engine/issues"
-  }, 
+  },
   "repository": {
-    "type": "git", 
+    "type": "git",
     "url": "https://github.com/playcanvas/engine.git"
-  }, 
+  },
   "eslintConfig": {
-    "extends": "@playcanvas/eslint-config", 
+    "extends": "@playcanvas/eslint-config",
     "parser": "babel-eslint"
-  }, 
+  },
   "eslintIgnore": [
-    "examples/lib/*", 
-    "scripts/textmesh/earcut.min.js", 
-    "scripts/textmesh/opentype.min.js", 
+    "examples/lib/*",
+    "scripts/textmesh/earcut.min.js",
+    "scripts/textmesh/opentype.min.js",
     "src/polyfill/*"
-  ], 
+  ],
   "files": [
-    "build/playcanvas.js", 
-    "build/playcanvas.min.js", 
-    "build/playcanvas.mjs", 
-    "build/playcanvas.dbg.js", 
-    "build/playcanvas.prf.js", 
-    "build/playcanvas.d.ts", 
-    "build/playcanvas-extras.js", 
-    "LICENSE", 
-    "package.json", 
-    "README.md", 
+    "build/playcanvas.js",
+    "build/playcanvas.min.js",
+    "build/playcanvas.mjs",
+    "build/playcanvas.dbg.js",
+    "build/playcanvas.prf.js",
+    "build/playcanvas.d.ts",
+    "build/playcanvas-extras.js",
+    "LICENSE",
+    "package.json",
+    "README.md",
     "README-zh.md"
-  ], 
+  ],
   "devDependencies": {
-    "@babel/core": "^7.14.5", 
-    "@babel/plugin-proposal-class-properties": "^7.14.5", 
-    "@babel/preset-env": "^7.14.5", 
-    "@playcanvas/eslint-config": "1.0.8", 
-    "@playcanvas/jsdoc-template": "1.0.19", 
-    "@rollup/plugin-babel": "5.3.0", 
-    "@rollup/plugin-replace": "2.4.2", 
-    "@rollup/pluginutils": "4.1.0", 
-    "babel-eslint": "10.1.0", 
-    "chai": "4.3.4", 
-    "eslint": "^7.28.0", 
-    "http-server": "0.12.3", 
-    "jsdoc": "^3.6.7", 
-    "karma": "^6.3.4", 
-    "karma-chrome-launcher": "3.1.0", 
-    "karma-mocha": "2.0.1", 
-    "karma-spec-reporter": "0.0.32", 
-    "mocha": "^9.0.0", 
-    "rollup": "^2.51.2", 
-    "rollup-plugin-jscc": "2.0.0", 
-    "rollup-plugin-terser": "7.0.2", 
-    "sinon": "^11.1.1", 
-    "tsd-jsdoc": "2.5.0", 
+    "@babel/core": "^7.14.5",
+    "@babel/plugin-proposal-class-properties": "^7.14.5",
+    "@babel/preset-env": "^7.14.5",
+    "@playcanvas/eslint-config": "1.0.8",
+    "@playcanvas/jsdoc-template": "1.0.19",
+    "@rollup/plugin-babel": "5.3.0",
+    "@rollup/plugin-replace": "2.4.2",
+    "@rollup/pluginutils": "4.1.0",
+    "babel-eslint": "10.1.0",
+    "chai": "4.3.4",
+    "eslint": "^7.28.0",
+    "http-server": "0.12.3",
+    "jsdoc": "^3.6.7",
+    "karma": "^6.3.4",
+    "karma-chrome-launcher": "3.1.0",
+    "karma-mocha": "2.0.1",
+    "karma-spec-reporter": "0.0.32",
+    "mocha": "^9.0.0",
+    "rollup": "^2.51.2",
+    "rollup-plugin-jscc": "2.0.0",
+    "rollup-plugin-terser": "7.0.2",
+    "sinon": "^11.1.1",
+    "tsd-jsdoc": "2.5.0",
     "typescript": "^4.3.2"
-  }, 
+  },
   "scripts": {
-    "build": "rollup -c", 
-    "build:es5": "rollup -c --environment target:es5", 
-    "build:es5min": "rollup -c --environment target:es5min", 
-    "build:es6": "rollup -c --environment target:es6", 
-    "build:debug": "rollup -c --environment target:debug", 
-    "build:profiler": "rollup -c --environment target:profiler", 
-    "docs": "jsdoc -c conf-api.json", 
-    "lint": "eslint --ext .js extras scripts src rollup.config.js", 
-    "serve": "http-server build -a localhost -p 51000", 
-    "test": "karma start tests/karma.conf.js -- --single-run --release", 
-    "test:watch": "karma start tests/karma.conf.js", 
-    "test:debug": "karma start tests/karma.conf.js -- --single-run=false", 
-    "test:tsd": "npm run tsd && tsc --pretty false build/playcanvas.d.ts", 
-    "tsd": "jsdoc -c conf-tsd.json && node tsd.js", 
+    "build": "rollup -c",
+    "build:es5": "rollup -c --environment target:es5",
+    "build:es5min": "rollup -c --environment target:es5min",
+    "build:es6": "rollup -c --environment target:es6",
+    "build:debug": "rollup -c --environment target:debug",
+    "build:profiler": "rollup -c --environment target:profiler",
+    "docs": "jsdoc -c conf-api.json",
+    "lint": "eslint --ext .js extras scripts src rollup.config.js",
+    "serve": "http-server build -a localhost -p 51000",
+    "test": "karma start tests/karma.conf.js -- --single-run --release",
+    "test:watch": "karma start tests/karma.conf.js",
+    "test:debug": "karma start tests/karma.conf.js -- --single-run=false",
+    "test:tsd": "npm run tsd && tsc --pretty false build/playcanvas.d.ts",
+    "tsd": "jsdoc -c conf-tsd.json && node tsd.js",
     "unzipbundle": "gzip -x tests/assets/bundle.gz"
   }
 }
